backend/requirements.txt
fastapi
uvicorn
python-multipart
pymongo
pdfplumber
python-dotenv
requests
huggingface_hub 